Move batch actions to separate file
