+++ src/components/Cluster/NewCluster/CreateNodePoolsCluster.js
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-import { batchedClusterCreate } from 'actions/batchedActions';
 import MasterNodes from 'Cluster/NewCluster/MasterNodes';
 import produce from 'immer';
 import PropTypes from 'prop-types';
@@ -7,6 +6,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { TransitionGroup } from 'react-transition-group';
 import { Constants } from 'shared/constants';
+import { batchedClusterCreate } from 'stores/batchActions';
 import { CLUSTER_CREATE_REQUEST } from 'stores/cluster/constants';
 import { selectErrorByAction } from 'stores/error/selectors';
 import SlideTransition from 'styles/transitions/SlideTransition';
