+++ src/components/Cluster/NewCluster/CreateRegularCluster.js
@@ -1,11 +1,11 @@
 import styled from '@emotion/styled';
-import { batchedClusterCreate } from 'actions/batchedActions';
 import InstanceTypeSelector from 'Cluster/ClusterDetail/InstanceTypeSelector/InstanceTypeSelector';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import { Constants, Providers } from 'shared/constants';
 import NodeCountSelector from 'shared/NodeCountSelector';
+import { batchedClusterCreate } from 'stores/batchActions';
 import { CLUSTER_CREATE_REQUEST } from 'stores/cluster/constants';
 import { selectErrorByAction } from 'stores/error/selectors';
 import Button from 'UI/Button';
